Add route handler for creating new pages
This will be the C in CRUD, essentially we are first checking if a valid login session exists, before accepting a POST request from the new page view and parsing the form data into a PHP array. We perform validation to make sure that at least the 'path' field is submitted so that we know where to save the newly created file. The rest of the function involves extracting and encoding the page metadata as JSON and writing it to the new .md file along with the page content.