Add utility classes for monospaced textfields, small text, and text submit buttons
Monospaed textfields: when we expect users to input code instead of literally text (e.g. in the footer, header, and snippet fields)

Small-text: for input field descriptions (so they don't take up so much space)

Text submit buttons: Used when we need a submit button with a name and value, but want it to appear as if it were an anchor insteasd (e.g. the 'X' link to delete media in the new/edit page views)