Handle the unlinking and writing of pages with validation
Specifically, not allow greater than 2 levels of nesting (no need for that headache yet), as well as handle failure cases with appropriate error messages.

We now check for and attempt to create a path structure if the relevant directories do not exist yet, so that the files don't disappear into thin air when we try writing them to nonexistent folders.